+++ example/lib/main.dart
@@ -8,6 +8,7 @@ import 'pages/point.dart';
 import 'pages/polygon.dart';
 import 'pages/custom.dart';
 import 'pages/bigdata.dart';
+import 'pages/echarts.dart';
 // import 'pages/debug.dart';
 
 final routes = {
@@ -19,6 +20,7 @@ final routes = {
   '/examples/Polygon Element': (context) => PolygonPage(),
   '/examples/Custom': (context) => CustomPage(),
   '/examples/Bigdata': (context) => BigdataPage(),
+  '/examples/Echarts': (context) => EchartsPage(),
   // '/examples/Debug': (context) => DebugPage(),
 };